Add part access by cpid
