+++ tests/odm/test_core.py
@@ -3,45 +3,34 @@
 import pytest
 
 # Local Modules
-from cofactr.graph import GraphAPI
-from cofactr.odm.part import Part
-
-
-from typing import List
-
-
-def search_parts(query: str, limit=10, external=True) -> List[Part]:
-    """Search for parts."""
-
-    graph = GraphAPI()
-
-    cursor = graph.get_products(
-        query=query,
-        fields="id,statements{spec,image,mfr,mpn,desc,package,numberofpins},offers",
-        batch_size=100,
-        limit=limit,
-        external=external,
-        render=False,
-    )
-
-    return [Part(**data) for data in cursor]
+from cofactr.odm.core import get_part, search_parts
 
 
 @pytest.mark.parametrize(
     "mpn",
     ["IRFH4251DTRPBF"],
 )
-def test_part(mpn: str):
+def test_search_part(mpn: str):
     """Test Part."""
 
     parts = search_parts(query=mpn, limit=1, external=False)
 
     assert len(parts) == 1
 
-    part = parts[0]
+
+@pytest.mark.parametrize(
+    "cpid",
+    ["TRGC72NRRA4W"],
+)
+def test_get_part(cpid: str):
+    """Test Part."""
+
+    part = get_part(cpid=cpid, external=False)
+
+    assert part
 
     assert part.id == "TRGC72NRRA4W"
-    assert part.mpn == mpn
+    assert part.mpn == "IRFH4251DTRPBF"
     assert part.hero_image == "https://assets.cofactr.com/TRGC72NRRA4W/part-img.jpg"
     assert (
         part.description
